Fixed bug where functions weren't called correctly
It turns out that the and operator is only for logic and cannot be used to call two different functions as I had intended. Thus I created a helper function that is called by the button on command and this helper function calls the original function as well as the destroy function. I also created a few placeholder functions that will be filled out later such as pauseInterval and resetInterval.